[DeepGemm]: optimize mqa_logits
1. use uint32_t for some stride, reduce IALU pressure
2. fp4 bf16 logits use tb_per_cu=2 tile

to #79982600: mqa_logits
Test Case: test_attention.py
Reviewed By: Shen Zhenghai

(cherry picked from commit 7c42fa525c83ad4b91fedbd7140d991a710d19cd)
(cherry picked from commit 17d894eef15587fa9f46ab2ee29d632aba00ce7a)

+++ deep_gemm/include/deep_gemm/fp4_paged_mqa_logits.cuh
@@ -185,7 +185,7 @@ public:
         const uint32_t batch_size;
         const uint64_t logits_stride;
         const uint64_t kv_cache_stride_bytes;
-        const uint64_t block_table_stride;
+        const uint32_t block_table_stride;
         const uint32_t* context_lens;
         ElementLogits* logits;
         const uint32_t* block_table;
@@ -614,7 +614,7 @@ class PagedAttentionFP4 {
 public:
     static void run(const ElementQK* ptr_q, const uint32_t* q_sf, const ElementQK* ptr_k, const uint32_t* k_sf,
                     const float* weights, const uint32_t batch_size, const uint64_t logits_stride,
-                    const uint64_t kv_cache_stride_bytes, const uint64_t block_table_stride,
+                    const uint64_t kv_cache_stride_bytes, const uint32_t block_table_stride,
                     const uint32_t* context_lens, ElementLogits* logits, const uint32_t* block_table,
                     const uint32_t* schedule_meta, cudaStream_t stream, int num_sms, int num_blocks) {
         using AttnKernel =
